fix: :bug: fix npm bump again

+++ bin/pre-push.ts
@@ -37,6 +37,13 @@ if (process.env[INNER_PRE_HOOK]) {
   process.exit(0)
 }
 
+const packageVersion = require('../package.json').version
+const lastCommitMsg = shell.exec('git log --pretty=format:"%s" HEAD^0 -1', { silent : true }).stdout
+
+if (packageVersion === lastCommitMsg) {
+  process.exit(0)
+}
+
 shell.exec('npm run lint').code === 0 || process.exit(1)
 shell.rm('-f', 'package-lock.json')
 shell.exec('npm version patch --no-package-lock').code === 0 || process.exit(1)
